move data_gen for tests to test folder
